restructure mcp
