+++ packages/mcp/src/prompts/applySemanticLayout.ts
@@ -1,0 +1,59 @@
+import type { ProjectId } from '@likec4/core'
+import type { LikeC4LanguageServices } from '@likec4/language-server'
+import { completable } from '@modelcontextprotocol/sdk/server/completable'
+import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
+import { prop } from 'remeda'
+import * as z from 'zod/v3'
+
+export function registerApplySemanticLayoutPrompt(
+  mcp: McpServer,
+  services: LikeC4LanguageServices,
+): McpServer {
+  mcp.registerPrompt(
+    'apply_semantic_layout',
+    {
+      title: 'Apply semantic layout to a view',
+      argsSchema: {
+        projectId: completable(
+          z.string()
+            // .default('default' as ProjectId)
+            .describe('Project id (optional, will use "default" if not specified)'),
+          (value = '') => {
+            return services.projects().filter(project => project.id.startsWith(value)).map(prop('id'))
+          },
+        ),
+        viewId: completable(
+          z.string().describe('View ID to apply semantic layout to'),
+          async (value = '', ctx) => {
+            const projectId = ctx?.arguments?.['projectId'] ?? services.projectsManager.default.id
+            const model = await services.computedModel(projectId as ProjectId)
+            if (!model) {
+              await mcp.sendLoggingMessage({
+                level: 'warning',
+                data: `Completing viewId for project ${projectId} with input value "${value}", model not found`,
+              })
+
+              return []
+            }
+            return Array.from(model.views()).filter((view) => view.id.startsWith(value)).map(prop('id'))
+          },
+        ),
+      },
+    },
+    async ({ viewId, projectId }, _ctx) => {
+      return {
+        messages: [{
+          role: 'user',
+          content: {
+            type: 'text',
+            text: `Call \`apply-semantic-layout\` tool of likec4 mcp with these arguments:\n${
+              JSON.stringify({ viewId, projectId })
+            }`,
+          },
+        }],
+      }
+    },
+  )
+
+  return mcp
+}
